requirements.txt
# requirements for full MNE-Python functionality plus kitchen sink 02.25.22
autopep8
autoreject
black
coloredlogs
datalad
decorator
dipy
docutils<0.18,>=0.14
fire
flake8
git+https://github.com/tomv564/pyls-mypy.git
h5io
h5py
imageio-ffmpeg>=0.4.1
imageio>=2.6.1
ipykernel
ipympl
ipyvtklink
ipywidgets
ipywidgets
jinja2
joblib
json_tricks
jupyter
markupsafe~=2.0.1
matplotlib
mccabe<0.7,>=0.6
mffpy>=0.5.7
mistune<2,>=0.8.1
mne-qt-browser
nibabel
nilearn
numba
numexpr
numpy<1.22,>=1.18
numpydoc
openpyxl
packaging
pandas
pandas-flavor
pandas-profiling
pooch>=1.5
psutil
pyjanitor
pylint
pymatreader
pyobjc-framework-Cocoa>=5.2.0; platform_system=="Darwin"
pyqt5-sip
python-language-server[all]
python-lsp-server[all]
python-picard
pyvista>=0.32
pyvistaqt>=0.4
scikit-learn
scipy>=1.1.0
seaborn
sip
sphinx
sphinx_bootstrap_theme
sphinx_fontawesome
statsmodels
tqdm
traitlets
xarray
xlrd
yapf
